+++ internal/app/gophermartprovider/gophermart_models.go
@@ -24,5 +24,5 @@ type GophermartProvider interface {
 	// GetOrders - берем все заказы пользователя из нашей базы данных
 	GetOrders(ctx context.Context, tx pgclient.Transaction, userID string) ([]models.Order, error)
 	// ExpandOrder - добавить информацию о заказе из accrual сервиса
-	ExpandOrder(ctx context.Context, tx pgclient.Transaction, status string, accrual *uint32, userID string) error
+	ExpandOrder(ctx context.Context, tx pgclient.Transaction, status string, accrual *float32, userID string) error
 }
